[client] Make redis_client_command_finish() return a GVariant.
We transform the redis reply to a variant in this patch. It allows
for clean binding to higher level languages as well as provides
clearer type information while in C. It does come with a slight
cost though, being the overhead of GVariant and copying the reply
into structures. That is minimal compared to anything custom we
would have come up with to copy the reply into a reference counted
structure.